Fix code smells

+++ src/libsync/theme.h
@@ -474,7 +474,7 @@ public:
     */
 	virtual QIcon uiThemeIcon(const QString &iconName, bool uiHasDarkBg) const;
 
-    Q_INVOKABLE static float getColorDarkness(const QColor &color);
+    Q_INVOKABLE static double getColorDarkness(const QColor &color);
 
     /**
      * @brief Perform a calculation to check if a colour is dark or light and accounts for different sensitivity of the human eye.
